I don't understand how this script should be used...
When I just call the script gcc-rtags-wrapper without any arguments it
will always land in the else branch (line 81), which will lead sooner or
later to a broken PATH variable...BUG?

The if branch (line 51) will only be entered, if the script name is
different than gcc-rtags-wrapper..., this means either a symlink or a
copy to a different name (mpicc/mpicxx/SOMETHING) must be created. Which
one? Symlink won't work due to the symlink resolving (line 40).
If a copy must be made, than in my opinion, it shouldn't be installed
into the bin directory, rather, in a configuration place. Don't know
which path this should be, right now, and to user must create than a
copy of it with a different name to a path which is in the env PATH
variable. But this should be properly documented in the shell script.

Also I don't really think someone uses it, according to the commit
history it was created a long time ago.

What do you think Anders, should we remove it entirely, create something
new...?